Set the fill_upsampling_holes parameter to True
Due to the resolution disparity between the depth images and the rgb
images from the embedded D410 rgbd camera on the gen3 kinova, when the
depth images are registered in the color image frame, many holes occur
in the registered depth data. Turning on fill_upsampling_holes spreads
the lower resolution depth data out across the registered image to
eliminate these holes.